Address timeout on refresh with 'cached' responses.
To handle the case where a refresh token is used to generate a new access token but the new access/refresh token is never received, allow a refresh token to be re-used within a grace period, returning the same access token (appearing as if the request/response was 'cached').

This now marks the RefreshToken revoked with a timestamp, instead of deleting it.

Ref: https://github.com/evonove/django-oauth-toolkit/issues/497