Run black over the source tree
This is the result of running `black .` over the repository.

By-hand improvements of the blackened code will be in follow up commits,
to make it easier to reapply this commit to future updates, if
necessary - IE to remove this commit and re-run black over a fresh tree,
rather than trying to merge new changes in to this commit.