fix: npm install failing
it seems we merged a major update from dependabot to some of the RP
dependencies that were incompatible with some of the other package
versions. This resolves that and updates the RP to Svelte4/SvelteKit 2.

Test by following the instructions in /test/app/README.md to launch
the local environment.