Honor database assignment from router (#1450)
* Improve multiple database support.

The token models might not be stored in the default database. There might not _be_ a default database.
Intead, the code now relies on Django's routers to determine the actual database to use when creating
transactions. This required moving from decorators to context managers for those transactions.

To test the multiple database scenario a new settings file as added which derives from settings.py
and then defines different databases and the routers needed to access them. The commit is larger than
might be expected because when there are multiple databases the Django tests have to be told which
databases to work on. Rather than copying the various test cases or making multiple database specific
ones the decision was made to add wrappers around the standard Django TestCase classes and programmatically
define the databases for them. This enables all of the same test code to work for both the one database
and the multi database scenarios with minimal maintenance costs. A tox environment that uses the
multi db settings file has been added to ensure both scenarios are always tested.

* changelog entry and authors update

* PR review response.

Document multiple database requires in advanced_topics.rst.
Add an ImproperlyConfigured validator to the ready method of the DOTConfig app.
Fix IDToken doc string.
Document the use of _save_bearer_token.
Define LocalIDToken and use it for validating the configuration test.
Questionably, define py39-multi-db-invalid-token-configuration-dj42. This will
consistently cause tox runs to fail until it is worked out how to mark this
as an expected failure.

* move migration

* update migration

* use django checks system

* drop misconfigured db check. Let's find a better way.

* run checks

* maybe a better test definition

* listing tests was breaking things

* No more magic.

* Oops. Debugger.

* Use retrieven_current_databases in django_db marked tests.

* Updates.

Prove the checks work.
Document test requirements.

* fix typo

---------

Co-authored-by: Alan Crosswell <alan@crosswell.us>
Co-authored-by: Alan Crosswell <alan@columbia.edu>