Make user field on abstract refresh token nullable
why?

DOT currently assume the user will be derived from the django request.user object (from the logic throughout DOT, not the model itself).

Since the device flow happens out of band there is no request.user available when the call to token is made, we have to make this field none.

How do I handle it in my own custom auth server:
In my custom auth server how I associate a refresh token with a user is to have a field (column in the refresh token table) that has the payload of the original JWT what was made when the refresh token was issued and I use the sub claim in the payload to know “this user has the refresh token” which prevents it relying on django solely for the user information but the stateless JWT instead