Improve settings documentation
Previously, it was unclear that the swappable model settings
should always be set without a namespace, as the sub-section
titles didn't include the required `OAUTH2_PROVIDER_` prefix.
The warning at the top may not be noticed by people looking for
a specific settings, and it was still unclear given the sub-section
titles.

Additionally, this documents the
`OAUTH2_PROVIDER_ID_TOKEN_MODEL` and
`OAUTH2_PROVIDER_GRANT_MODEL` settings, which
were previously undocumented.

Also, this corrects a mistake where `ACCESS_TOKEN_GENERATOR` was
mentioned as the setting which enables the use of `SettingsScopes`; the
actual setting is `SCOPES_BACKEND_CLASS`.