fix: addressed comments; enhanced preconditions and responses of views
In this commit I've addressed a few issues raised in comments about
the flow.

* Always return informative errors to user-facing views. If the device
is not in the expected state, the errors are returned in the form, instead
of raising exceptions.

* Always return JSON response to device-facing view (aka TokenView). If
the device is not in the expected state, the errors are returned according
to the RFC.

* Never involve device_code in frontend. The redirect to device-confirm
view now takes client_id and user_code arguments instead of device_code

* Increased test coverage. Added tests for expected error cases handled
in the code